Sage: - fixed msi repeat selection - lower phasing read count for sbx and ultima - various ultima fixes  various ultima fixes

+++ sage/src/main/java/com/hartwig/hmftools/sage/SageConstants.java
@@ -128,7 +128,7 @@ public class SageConstants
     // quality calcs
     public static final int DEFAULT_JITTER_MIN_REPEAT_COUNT = 3;
     public static final double JITTER_QUAL_BOOST_MAX_PERC = 1.3;
-    public static final int DEFAULT_BASE_QUAL_FIXED_PENALTY = 12;
+    public static final byte DEFAULT_BASE_QUAL_FIXED_PENALTY = 12;
 
     public static final int READ_EDGE_PENALTY_0 = 15;
     public static final int READ_EDGE_PENALTY_1 = 5;
